remote TelemetryContext from ApplicationAnalytics package
